Refactoring fetching games using ReactQuery
