--- src/hooks/useData.ts
@@ -1,38 +1,0 @@
-import { useEffect, useState } from "react";
-import apiClient from "../services/api-client";
-import { AxiosRequestConfig, CanceledError } from "axios";
-
-export interface FetchResponse<T> {
-    count: number;
-    results: T[];
-
-}
-
-const useData = <T>(endpoint: string, requestConfig?: AxiosRequestConfig, deps?: any[]) => {
-          const [data, setData] = useState<T[]>([]);
-          const [error, setError] = useState("");
-          const [isLoading, setLoading] = useState(false);
-        
-          useEffect(() => {
-            const controller = new AbortController();
-    
-            setLoading(true);
-            apiClient
-              .get<FetchResponse<T>>(endpoint, {signal: controller.signal, ...requestConfig})
-              .then((response) => {
-                setData(response.data.results);
-                setLoading(false);
-            })
-              .catch((err) => {
-                if (err instanceof CanceledError) return;
-                setError(err.message)
-                setLoading(false);
-            });
-    
-              return () => controller.abort();
-          }, deps ? [...deps] : []);
-    
-          return {data, error, isLoading}
-    
-}
-export default useData;
