+++ src/hooks/useGames.ts
@@ -1,6 +1,8 @@
+import { useQuery } from "@tanstack/react-query";
 import { GameQuery } from "../App";
-import useData from "./useData";
-import { Genre } from "./useGenres";
+import apiClient from "../services/api-client";
+import { FetchResponse } from "../services/api-client";
+
 
 export interface Platform {
     id: number;
@@ -17,13 +19,18 @@ export interface Game {
   rating_top: number;
 }
 
-const useGames = (gameQuery: GameQuery) => useData<Game>("/games", {
-   params: { 
-    genres: gameQuery?.genre?.id, 
-    platforms: gameQuery?.platform?.id,
-    ordering: gameQuery?.sortOrder,
-    search: gameQuery?.searchText
-  }}, 
-  [gameQuery]);
+const useGames = (gameQuery: GameQuery) => 
+  useQuery<FetchResponse<Game>, Error>({
+    queryKey: ['games', gameQuery],
+    queryFn: () => apiClient.get<FetchResponse<Game>>('/games', {
+      params: { 
+        genres: gameQuery?.genre?.id, 
+        parent_platforms: gameQuery?.platform?.id,
+        ordering: gameQuery?.sortOrder,
+        search: gameQuery?.searchText
+      }
+    }).then(res => res.data),
+    staleTime: 24 * 60 * 60 * 1000, // 24h
+  })
 
 export default useGames;
