+++ src/hooks/usePlatforms.ts
@@ -1,6 +1,6 @@
 import { useQuery } from "@tanstack/react-query";
-import useData, { FetchResponse } from "./useData";
 import apiClient from "../services/api-client";
+import { FetchResponse } from "../services/api-client";
 import { Platform } from "./useGames";
 
 interface Plaforms {
