+++ src/services/api-client.ts
@@ -1,5 +1,11 @@
 import axios from "axios";
 
+export interface FetchResponse<T> {
+    count: number;
+    results: T[];
+}
+
+
 export default axios.create({
     baseURL:"https://api.rawg.io/api",
     params: {
